+++ implants/imix/src/tasks.rs
@@ -50,7 +50,7 @@ pub async fn get_new_tasks(
 
 pub async fn start_new_tasks(
     new_tasks: Vec<Task>,
-    all_exec_futures: &mut HashMap<i32, AsyncTask>,
+    all_exec_futures: &mut HashMap<i64, AsyncTask>,
     debug_start_time: Instant,
 ) -> Result<()> {
     for task in new_tasks {
